ImageViewer: add per-channel mask toggle.
We could also request a subset of the channels of the current group when fetching the texture to display, but this would pack them (so that e.g. asking for "Green, Alpha" would result in a red & green image. We would have to pad with empty channels. We don't apply the channel mask on the CPU picker/labels for now.